# specs/001-todo-rest-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo REST API
  description: |
    Multi-user RESTful API for todo management with JWT-based authentication and user-level data isolation.

    ## Authentication
    All endpoints require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## User Isolation
    All operations are scoped to the authenticated user. Users can only access, modify, or delete their own todos.

    ## Error Responses
    All error responses follow this format:
    ```json
    {
      "detail": "Human-readable error message",
      "error_code": "MACHINE_READABLE_CODE"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server

tags:
  - name: todos
    description: Todo management operations

security:
  - BearerAuth: []

paths:
  /todos:
    get:
      tags:
        - todos
      summary: List todos for authenticated user
      description: Retrieve a paginated list of todos owned by the authenticated user, optionally filtered by completion status.
      operationId: listTodos
      parameters:
        - name: skip
          in: query
          description: Number of todos to skip (for pagination)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of todos to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: completed
          in: query
          description: Filter by completion status (omit for all todos)
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response with paginated todo list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
              examples:
                allTodos:
                  summary: All todos
                  value:
                    items:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        title: Buy groceries
                        description: Milk, eggs, bread
                        completed: false
                        user_id: user_123
                        created_at: '2026-02-13T10:00:00Z'
                        updated_at: '2026-02-13T10:00:00Z'
                        completed_at: null
                      - id: 660e8400-e29b-41d4-a716-446655440001
                        title: Finish project
                        description: Complete Phase II implementation
                        completed: true
                        user_id: user_123
                        created_at: '2026-02-12T09:00:00Z'
                        updated_at: '2026-02-13T11:00:00Z'
                        completed_at: '2026-02-13T11:00:00Z'
                    total: 2
                    skip: 0
                    limit: 20
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - todos
      summary: Create a new todo
      description: Create a new todo for the authenticated user.
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
            examples:
              minimal:
                summary: Minimal todo (title only)
                value:
                  title: Buy groceries
              withDescription:
                summary: Todo with description
                value:
                  title: Finish project
                  description: Complete Phase II implementation
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                title: Buy groceries
                description: null
                completed: false
                user_id: user_123
                created_at: '2026-02-13T10:00:00Z'
                updated_at: '2026-02-13T10:00:00Z'
                completed_at: null
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /todos/{todo_id}:
    get:
      tags:
        - todos
      summary: Get a specific todo
      description: Retrieve a single todo by its ID. User must own the todo.
      operationId: getTodo
      parameters:
        - name: todo_id
          in: path
          description: UUID of the todo to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                title: Buy groceries
                description: Milk, eggs, bread
                completed: false
                user_id: user_123
                created_at: '2026-02-13T10:00:00Z'
                updated_at: '2026-02-13T10:00:00Z'
                completed_at: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - todos
      summary: Update a todo
      description: Update title, description, or completion status of a todo. User must own the todo. All fields are optional.
      operationId: updateTodo
      parameters:
        - name: todo_id
          in: path
          description: UUID of the todo to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
            examples:
              updateTitle:
                summary: Update title only
                value:
                  title: Buy groceries and supplies
              markComplete:
                summary: Mark as complete
                value:
                  completed: true
              updateAll:
                summary: Update multiple fields
                value:
                  title: Finish project
                  description: Complete Phase II implementation and testing
                  completed: true
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                title: Buy groceries and supplies
                description: Milk, eggs, bread
                completed: false
                user_id: user_123
                created_at: '2026-02-13T10:00:00Z'
                updated_at: '2026-02-13T10:30:00Z'
                completed_at: null
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - todos
      summary: Delete a todo
      description: Permanently delete a todo. User must own the todo. This operation cannot be undone.
      operationId: deleteTodo
      parameters:
        - name: todo_id
          in: path
          description: UUID of the todo to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Better Auth

  schemas:
    TodoCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Todo title (required, cannot be empty)
          example: Buy groceries
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Optional detailed description
          example: Milk, eggs, bread

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
          description: Updated todo title
          example: Buy groceries and supplies
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Updated description
          example: Milk, eggs, bread, coffee
        completed:
          type: boolean
          nullable: true
          description: Updated completion status
          example: true

    TodoResponse:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the todo
          example: 550e8400-e29b-41d4-a716-446655440000
        title:
          type: string
          maxLength: 200
          description: Todo title
          example: Buy groceries
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Todo description
          example: Milk, eggs, bread
        completed:
          type: boolean
          description: Completion status
          example: false
        user_id:
          type: string
          description: Owner's user identifier
          example: user_123
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: '2026-02-13T10:00:00Z'
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: '2026-02-13T10:00:00Z'
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp (UTC, null if not completed)
          example: null

    TodoListResponse:
      type: object
      required:
        - items
        - total
        - skip
        - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
          description: List of todos
        total:
          type: integer
          minimum: 0
          description: Total number of todos matching the filter
          example: 42
        skip:
          type: integer
          minimum: 0
          description: Number of todos skipped
          example: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Maximum number of todos returned
          example: 20

    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: Todo not found
        error_code:
          type: string
          description: Machine-readable error code
          example: TODO_NOT_FOUND

  responses:
    BadRequestError:
      description: Invalid request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Title cannot be empty
            error_code: VALIDATION_ERROR

    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Invalid or expired JWT token
            error_code: UNAUTHORIZED

    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: You do not have permission to access this todo
            error_code: FORBIDDEN

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: Todo not found
            error_code: TODO_NOT_FOUND

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: An unexpected error occurred
            error_code: INTERNAL_SERVER_ERROR